[E2E] Refactor: Separate codebase into new files (#261)
