+++ test/e2e/e2e_ova_from_url_test.go
@@ -3,6 +3,10 @@ package e2e_test
 import (
 	"fmt"
 	"github.com/kubev2v/migration-planner/api/v1alpha1"
+	. "github.com/kubev2v/migration-planner/test/e2e"
+	. "github.com/kubev2v/migration-planner/test/e2e/e2e_agent"
+	. "github.com/kubev2v/migration-planner/test/e2e/e2e_helpers"
+	. "github.com/kubev2v/migration-planner/test/e2e/e2e_service"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"go.uber.org/zap"
@@ -23,8 +27,8 @@ var _ = Describe("e2e-download-ova-from-url", func() {
 
 	BeforeEach(func() {
 		startTime = time.Now()
-		testOptions.downloadImageByUrl = true
-		testOptions.disconnectedEnvironment = false
+		TestOptions.DownloadImageByUrl = true
+		TestOptions.DisconnectedEnvironment = false
 
 		svc, err = NewPlannerService()
 		Expect(err).To(BeNil(), "Failed to create PlannerService")
@@ -33,22 +37,26 @@ var _ = Describe("e2e-download-ova-from-url", func() {
 		Expect(err).To(BeNil())
 		Expect(source).NotTo(BeNil())
 
-		agent, err = CreateAgent(defaultAgentTestID, source.Id, vmName)
+		agent, err = CreateAgent(DefaultAgentTestID, source.Id, VmName)
 		Expect(err).To(BeNil())
 
 		zap.S().Info("Waiting for agent IP...")
 		Eventually(func() error {
-			return FindAgentIp(agent, &agentIP)
+			agentIP, err = agent.GetIp()
+			if err != nil {
+				return err
+			}
+			return nil
 		}, "3m", "2s").Should(BeNil())
 		zap.S().Infof("Agent ip is: %s", agentIP)
 
 		agentApiBaseUrl := fmt.Sprintf("https://%s:3333/api/v1/", agentIP)
-		agent.AgentApi().baseURL = agentApiBaseUrl
+		agent.SetAgentApi(DefaultAgentApi(agentApiBaseUrl))
 		zap.S().Infof("Agent Api base url: %s", agentApiBaseUrl)
 
 		zap.S().Info("Wait for planner-agent to be running...")
 		Eventually(func() bool {
-			return IsPlannerAgentRunning(agent, agentIP)
+			return agent.IsServiceRunning(agentIP, "planner-agent")
 		}, "3m", "2s").Should(BeTrue())
 		zap.S().Info("Planner-agent is now running")
 
@@ -67,7 +75,7 @@ var _ = Describe("e2e-download-ova-from-url", func() {
 		Expect(err).To(BeNil(), "Failed to remove vm and iso")
 		testDuration := time.Since(startTime)
 		zap.S().Infof("Test completed in: %s\n", testDuration.String())
-		testsExecutionTime[CurrentSpecReport().LeafNodeText] = testDuration
+		TestsExecutionTime[CurrentSpecReport().LeafNodeText] = testDuration
 	})
 
 	AfterFailed(func() {
@@ -78,7 +86,7 @@ var _ = Describe("e2e-download-ova-from-url", func() {
 		It("Downloads OVA file from URL", func() {
 			zap.S().Infof("============Running test: %s============", CurrentSpecReport().LeafNodeText)
 
-			res, err := agent.AgentApi().Login(fmt.Sprintf("https://%s:%s/sdk", systemIP, Vsphere1Port),
+			res, err := agent.AgentApi().Login(fmt.Sprintf("https://%s:%s/sdk", SystemIP, Vsphere1Port),
 				"core", "123456")
 			Expect(err).To(BeNil())
 			Expect(res.StatusCode).To(Equal(http.StatusNoContent))
