+++ test/e2e/e2e_service/service.go
@@ -1,38 +1,28 @@
-package e2e_test
+package e2e_service
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"github.com/google/uuid"
 	"github.com/kubev2v/migration-planner/api/v1alpha1"
 	api "github.com/kubev2v/migration-planner/api/v1alpha1"
 	internalclient "github.com/kubev2v/migration-planner/internal/api/client"
+	. "github.com/kubev2v/migration-planner/test/e2e"
+	. "github.com/kubev2v/migration-planner/test/e2e/e2e_utils"
 	"go.uber.org/zap"
 	"net/http"
-	"os"
-	"strings"
 )
 
 type PlannerService interface {
-	RemoveSources() error
-	RemoveSource(id uuid.UUID) error
-	GetSource(id uuid.UUID) (*api.Source, error)
 	CreateSource(name string) (*api.Source, error)
+	GetSource(id uuid.UUID) (*api.Source, error)
+	RemoveSource(id uuid.UUID) error
+	RemoveSources() error
 	UpdateSource(uuid.UUID, *v1alpha1.Inventory) error
 }
 
-var (
-	defaultServiceUrl string = fmt.Sprintf("http://%s:3443", os.Getenv("PLANNER_IP"))
-)
-
 type plannerService struct {
 	api *ServiceApi
-}
-
-type ServiceApi struct {
-	baseURL    string
-	httpClient *http.Client
 }
 
 func NewPlannerService() (*plannerService, error) {
@@ -42,14 +32,14 @@ func NewPlannerService() (*plannerService, error) {
 
 func (s *plannerService) CreateSource(name string) (*api.Source, error) {
 	zap.S().Info("Creating source...")
-	user, err := defaultUserAuth()
+	user, err := DefaultUserAuth()
 	if err != nil {
 		return nil, err
 	}
 
 	params := &v1alpha1.CreateSourceJSONRequestBody{Name: name}
 
-	if testOptions.disconnectedEnvironment { // make the service unreachable
+	if TestOptions.DisconnectedEnvironment { // make the service unreachable
 
 		toStrPtr := func(s string) *string {
 			return &s
@@ -67,7 +57,7 @@ func (s *plannerService) CreateSource(name string) (*api.Source, error) {
 		return nil, err
 	}
 
-	res, err := s.api.request(http.MethodPost, "", reqBody, user.Token.Raw)
+	res, err := s.api.PostRequest("", reqBody, user.Token.Raw)
 	if err != nil {
 		return nil, err
 	}
@@ -88,12 +78,12 @@ func (s *plannerService) CreateSource(name string) (*api.Source, error) {
 }
 
 func (s *plannerService) GetSource(id uuid.UUID) (*api.Source, error) {
-	user, err := defaultUserAuth()
+	user, err := DefaultUserAuth()
 	if err != nil {
 		return nil, err
 	}
 
-	res, err := s.api.request(http.MethodGet, id.String(), nil, user.Token.Raw)
+	res, err := s.api.GetRequest(id.String(), user.Token.Raw)
 	if err != nil {
 		return nil, err
 	}
@@ -107,32 +97,13 @@ func (s *plannerService) GetSource(id uuid.UUID) (*api.Source, error) {
 	return getSourceRes.JSON200, nil
 }
 
-func (s *plannerService) RemoveSources() error {
-	user, err := defaultUserAuth()
-	if err != nil {
-		return err
-	}
-
-	res, err := s.api.request(http.MethodDelete, "", nil, user.Token.Raw)
-	if err != nil {
-		return err
-	}
-	defer res.Body.Close()
-
-	if res.StatusCode != http.StatusOK {
-		return fmt.Errorf("failed to delete sources. response status code: %d", res.StatusCode)
-	}
-
-	return err
-}
-
 func (s *plannerService) RemoveSource(uuid uuid.UUID) error {
-	user, err := defaultUserAuth()
+	user, err := DefaultUserAuth()
 	if err != nil {
 		return err
 	}
 
-	res, err := s.api.request(http.MethodDelete, uuid.String(), nil, user.Token.Raw)
+	res, err := s.api.DeleteRequest(uuid.String(), user.Token.Raw)
 	if err != nil {
 		return err
 	}
@@ -146,8 +117,27 @@ func (s *plannerService) RemoveSource(uuid uuid.UUID) error {
 	return err
 }
 
+func (s *plannerService) RemoveSources() error {
+	user, err := DefaultUserAuth()
+	if err != nil {
+		return err
+	}
+
+	res, err := s.api.DeleteRequest("", user.Token.Raw)
+	if err != nil {
+		return err
+	}
+	defer res.Body.Close()
+
+	if res.StatusCode != http.StatusOK {
+		return fmt.Errorf("failed to delete sources. response status code: %d", res.StatusCode)
+	}
+
+	return err
+}
+
 func (s *plannerService) UpdateSource(uuid uuid.UUID, inventory *v1alpha1.Inventory) error {
-	user, err := defaultUserAuth()
+	user, err := DefaultUserAuth()
 	if err != nil {
 		return err
 	}
@@ -162,7 +152,7 @@ func (s *plannerService) UpdateSource(uuid uuid.UUID, inventory *v1alpha1.Invent
 		return err
 	}
 
-	res, err := s.api.request(http.MethodPut, uuid.String(), reqBody, user.Token.Raw)
+	res, err := s.api.PutRequest(uuid.String(), reqBody, user.Token.Raw)
 	if err != nil {
 		return err
 	}
@@ -174,47 +164,4 @@ func (s *plannerService) UpdateSource(uuid uuid.UUID, inventory *v1alpha1.Invent
 	}
 
 	return err
-}
-
-func NewServiceApi() *ServiceApi {
-	return &ServiceApi{
-		baseURL:    fmt.Sprintf("%s/api/v1/sources/", defaultServiceUrl),
-		httpClient: &http.Client{},
-	}
-}
-
-func (api *ServiceApi) request(method string, path string, body []byte, jwtToken string) (*http.Response, error) {
-	var req *http.Request
-	var err error
-
-	queryPath := strings.TrimRight(api.baseURL+path, "/")
-
-	switch method {
-	case http.MethodGet:
-		req, err = http.NewRequest(http.MethodGet, queryPath, nil)
-	case http.MethodPost:
-		req, err = http.NewRequest(http.MethodPost, queryPath, bytes.NewReader(body))
-	case http.MethodPut:
-		req, err = http.NewRequest(http.MethodPut, queryPath, bytes.NewReader(body))
-	case http.MethodDelete:
-		req, err = http.NewRequest(http.MethodDelete, queryPath, nil)
-	default:
-		return nil, fmt.Errorf("unsupported method: %s", method)
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", jwtToken))
-
-	zap.S().Infof("[Service-API] %s [Method: %s]", req.URL.String(), req.Method)
-
-	res, err := api.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
 }
